+++ crates/assembly-syntax/src/parser/token.rs
@@ -498,8 +498,12 @@ pub enum Token<'input> {
     Locaddr,
     LocLoad,
     LocLoadw,
+    LocLoadwBe,
+    LocLoadwLe,
     LocStore,
     LocStorew,
+    LocStorewBe,
+    LocStorewLe,
     Lt,
     Lte,
     Mem,
@@ -716,8 +720,12 @@ impl fmt::Display for Token<'_> {
             Token::Locaddr => write!(f, "locaddr"),
             Token::LocLoad => write!(f, "loc_load"),
             Token::LocLoadw => write!(f, "loc_loadw"),
+            Token::LocLoadwBe => write!(f, "loc_loadw_be"),
+            Token::LocLoadwLe => write!(f, "loc_loadw_le"),
             Token::LocStore => write!(f, "loc_store"),
             Token::LocStorew => write!(f, "loc_storew"),
+            Token::LocStorewBe => write!(f, "loc_storew_be"),
+            Token::LocStorewLe => write!(f, "loc_storew_le"),
             Token::Lt => write!(f, "lt"),
             Token::Lte => write!(f, "lte"),
             Token::Mem => write!(f, "mem"),
@@ -926,8 +934,12 @@ impl<'input> Token<'input> {
                 | Token::Locaddr
                 | Token::LocLoad
                 | Token::LocLoadw
+                | Token::LocLoadwBe
+                | Token::LocLoadwLe
                 | Token::LocStore
                 | Token::LocStorew
+                | Token::LocStorewBe
+                | Token::LocStorewLe
                 | Token::Lt
                 | Token::Lte
                 | Token::Mem
@@ -1118,8 +1130,12 @@ impl<'input> Token<'input> {
         ("locaddr", Token::Locaddr),
         ("loc_load", Token::LocLoad),
         ("loc_loadw", Token::LocLoadw),
+        ("loc_loadw_be", Token::LocLoadwBe),
+        ("loc_loadw_le", Token::LocLoadwLe),
         ("loc_store", Token::LocStore),
         ("loc_storew", Token::LocStorew),
+        ("loc_storew_be", Token::LocStorewBe),
+        ("loc_storew_le", Token::LocStorewLe),
         ("lt", Token::Lt),
         ("lte", Token::Lte),
         ("mem", Token::Mem),