+++ crates/test-utils/src/lib.rs
@@ -75,13 +75,13 @@ pub const U32_BOUND: u64 = u32::MAX as u64 + 1;
 /// A source code of the `truncate_stack` procedure.
 pub const TRUNCATE_STACK_PROC: &str = "
 proc.truncate_stack.4
-    loc_storew.0 dropw movupw.3
+    loc_storew_be.0 dropw movupw.3
     sdepth neq.16
     while.true
         dropw movupw.3
         sdepth neq.16
     end
-    loc_loadw.0
+    loc_loadw_be.0
 end
 ";