+++ crates/test-utils/src/test_builders.rs
@@ -19,13 +19,13 @@ macro_rules! build_op_test {
     ($op_str:expr) => {{
         let source = format!("
 proc.truncate_stack.4
-    loc_storew.0 dropw movupw.3
+    loc_storew_be.0 dropw movupw.3
     sdepth neq.16
     while.true
         dropw movupw.3
         sdepth neq.16
     end
-    loc_loadw.0
+    loc_loadw_be.0
 end
 
 begin {} exec.truncate_stack end",
@@ -36,13 +36,13 @@ begin {} exec.truncate_stack end",
     ($op_str:expr, $($tail:tt)+) => {{
         let source = format!("
 proc.truncate_stack.4
-    loc_storew.0 dropw movupw.3
+    loc_storew_be.0 dropw movupw.3
     sdepth neq.16
     while.true
         dropw movupw.3
         sdepth neq.16
     end
-    loc_loadw.0
+    loc_loadw_be.0
 end
 
 begin {} exec.truncate_stack end",