+++ stdlib/tests/math/secp256k1/group.rs
@@ -18,26 +18,26 @@ fn test_secp256k1_point_doubling(src: Point, dst: Point) {
     proc.point_doubling_test_wrapper.48
         # push X -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.0
+        loc_storew_be.0
         dropw
         push.{}.{}.{}.{}
-        loc_storew.4
+        loc_storew_be.4
         dropw
 
         # push Y -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.8
+        loc_storew_be.8
         dropw
         push.{}.{}.{}.{}
-        loc_storew.12
+        loc_storew_be.12
         dropw
 
         # push Z -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.16
+        loc_storew_be.16
         dropw
         push.{}.{}.{}.{}
-        loc_storew.20
+        loc_storew_be.20
         dropw
 
         # input/ output memory addresses for point doubling purpose
@@ -214,50 +214,50 @@ fn test_secp256k1_point_addition(src0: Point, src1: Point, dst: Point) {
     proc.point_addition_test_wrapper.72
         # push X1 -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.0
+        loc_storew_be.0
         dropw
         push.{}.{}.{}.{}
-        loc_storew.4
+        loc_storew_be.4
         dropw
 
         # push Y1 -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.8
+        loc_storew_be.8
         dropw
         push.{}.{}.{}.{}
-        loc_storew.12
+        loc_storew_be.12
         dropw
 
         # push Z1 -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.16
+        loc_storew_be.16
         dropw
         push.{}.{}.{}.{}
-        loc_storew.20
+        loc_storew_be.20
         dropw
 
         # push X2 -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.24
+        loc_storew_be.24
         dropw
         push.{}.{}.{}.{}
-        loc_storew.28
+        loc_storew_be.28
         dropw
 
         # push Y2 -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.32
+        loc_storew_be.32
         dropw
         push.{}.{}.{}.{}
-        loc_storew.36
+        loc_storew_be.36
         dropw
 
         # push Z2 -coordinate to memory
         push.{}.{}.{}.{}
-        loc_storew.40
+        loc_storew_be.40
         dropw
         push.{}.{}.{}.{}
-        loc_storew.44
+        loc_storew_be.44
         dropw
 
         # input/ output memory addresses for point doubling purpose
@@ -477,27 +477,27 @@ fn test_secp256k1_point_multiplication(src_point: Point, scalar: FieldElement, d
 
         # EC point
         push.{}.{}.{}.{}
-        loc_storew.0
+        loc_storew_be.0
         dropw
 
         push.{}.{}.{}.{}
-        loc_storew.4
+        loc_storew_be.4
         dropw
 
         push.{}.{}.{}.{}
-        loc_storew.8
+        loc_storew_be.8
         dropw
 
         push.{}.{}.{}.{}
-        loc_storew.12
+        loc_storew_be.12
         dropw
 
         push.{}.{}.{}.{}
-        loc_storew.16
+        loc_storew_be.16
         dropw
 
         push.{}.{}.{}.{}
-        loc_storew.20
+        loc_storew_be.20
         dropw
 
         locaddr.20